# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/classes/{classId}/assessments": {
      "get": {
        "operationId": "AssessmentsController_list",
        "parameters": [
          {
            "name": "classId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "HOMEWORK",
                "GRADED_QUIZ",
                "PRACTICE_QUIZ"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PUBLISHED",
                "CLOSED"
              ]
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      },
      "post": {
        "operationId": "AssessmentsController_create",
        "parameters": [
          {
            "name": "classId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertAssessmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}": {
      "get": {
        "operationId": "AssessmentsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "assessments"
        ]
      },
      "patch": {
        "operationId": "AssessmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertAssessmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/recipients/preview": {
      "get": {
        "operationId": "AssessmentsController_preview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/publish": {
      "post": {
        "operationId": "AssessmentsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/close": {
      "post": {
        "operationId": "AssessmentsController_close",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/extend-deadline": {
      "post": {
        "operationId": "AssessmentsController_extend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtendDeadlineDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/duplicate": {
      "post": {
        "operationId": "AssessmentsController_duplicate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/submission": {
      "get": {
        "operationId": "AssessmentsController_draft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "assessments"
        ]
      },
      "patch": {
        "operationId": "AssessmentsController_saveDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPayloadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/submission/submit": {
      "post": {
        "operationId": "AssessmentsController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitPayloadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/result": {
      "get": {
        "operationId": "AssessmentsController_ownResult",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/homework-attachment": {
      "post": {
        "operationId": "AssessmentsController_uploadAttachment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/submissions/{submissionId}/grade": {
      "post": {
        "operationId": "AssessmentsController_grade",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeHomeworkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/practice/start": {
      "post": {
        "operationId": "AssessmentsController_startPractice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/practice/{attemptId}": {
      "patch": {
        "operationId": "AssessmentsController_savePractice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attemptId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeDraftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      },
      "get": {
        "operationId": "AssessmentsController_practiceDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attemptId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/practice/{attemptId}/submit": {
      "post": {
        "operationId": "AssessmentsController_submitPractice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attemptId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeDraftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{id}/practice": {
      "get": {
        "operationId": "AssessmentsController_practiceHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "HOMEWORK",
                "GRADED_QUIZ",
                "PRACTICE_QUIZ"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PUBLISHED",
                "CLOSED"
              ]
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "assessments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lingua LMS Backend API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lingua.sid"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "QuestionOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "isCorrect": {
            "type": "boolean"
          },
          "position": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "text",
          "isCorrect",
          "position"
        ]
      },
      "AssessmentQuestionDto": {
        "type": "object",
        "properties": {
          "options": {
            "minItems": 2,
            "maxItems": 20,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionOptionDto"
            }
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "enum": [
              "SINGLE_CHOICE",
              "TRUE_FALSE"
            ],
            "type": "string"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "explanation": {
            "type": "string",
            "maxLength": 2000
          },
          "points": {
            "type": "number",
            "minimum": 0.01,
            "maximum": 10000
          },
          "position": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "options",
          "type",
          "prompt",
          "points",
          "position"
        ]
      },
      "UpsertAssessmentDto": {
        "type": "object",
        "properties": {
          "questions": {
            "default": [],
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssessmentQuestionDto"
            }
          },
          "kind": {
            "enum": [
              "HOMEWORK",
              "GRADED_QUIZ",
              "PRACTICE_QUIZ"
            ],
            "type": "string"
          },
          "audience": {
            "enum": [
              "WHOLE_CLASS",
              "SELECTED_GROUPS"
            ],
            "type": "string"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "instructions": {
            "type": "string",
            "maxLength": 5000
          },
          "maxPoints": {
            "type": "number",
            "minimum": 0.01
          },
          "dueAt": {
            "type": "string"
          },
          "allowLate": {
            "type": "object",
            "default": false
          },
          "groupIds": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "materialIds": {
            "default": [],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "questions",
          "kind",
          "audience",
          "title",
          "allowLate",
          "groupIds",
          "materialIds"
        ]
      },
      "ExtendDeadlineDto": {
        "type": "object",
        "properties": {
          "dueAt": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        },
        "required": [
          "dueAt",
          "reason"
        ]
      },
      "DraftPayloadDto": {
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "number",
            "minimum": 0
          },
          "textAnswer": {
            "type": "string",
            "maxLength": 20000
          },
          "answers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "attachmentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "expectedRevision"
        ]
      },
      "SubmitPayloadDto": {
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "number",
            "minimum": 0
          },
          "textAnswer": {
            "type": "string",
            "maxLength": 20000
          },
          "answers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "attachmentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "expectedRevision"
        ]
      },
      "GradeHomeworkDto": {
        "type": "object",
        "properties": {
          "earnedPoints": {
            "type": "number",
            "minimum": 0
          },
          "feedback": {
            "type": "string",
            "maxLength": 5000
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        },
        "required": [
          "earnedPoints"
        ]
      },
      "PracticeDraftDto": {
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "number",
            "minimum": 0
          },
          "answers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "expectedRevision",
          "answers"
        ]
      }
    }
  }
}